Make a specific function to build linest matrix on discrete factors
The function buildFactorMatrix create a matrix that correspond to
tX * X where X is the linear model matrix. It works in the case there
are only factors with discrete levels and use the countOccurrences
function. The second argument "factors" of the functions specify a list
of the factors that should be used to build the matrix.