Improve robustness over minor CSV variations
The X and Y coordinates were missing in some cases. To avoid problems
extract the measurements data by counting the number fields in the
headers.

Add a function to filter empty cells. Useful for headers because sometimes
an empty cell is present at the end on the row.

Perform also a tranformation of headers by removing surrounding spaces
and quotes. Sometimes the strings are quoted with " and sometimes not.